refactor(application): extract vulnerability_builder.rs [#314-4/5]
Move build_vulnerabilities, build_vulnerability_views_for_package,
build_vulnerability_view, and map_severity from mod.rs into a dedicated
vulnerability_builder.rs module.

- build_vulnerabilities and map_severity are pub(super) for sibling access
- build_vulnerability_views_for_package and build_vulnerability_view are private
- mod.rs delegates to vulnerability_builder:: free functions
- Tests updated to call vulnerability_builder:: functions directly

Closes #340
